+++ database/migrations/2026_01_22_235410_add_attendance_fields_to_meeting_attendances_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             
             // Add foreign key columns
             $table->unsignedInteger('meeting_id')->after('attendance_id');
-            $table->unsignedInteger('student_id')->after('meeting_id');
+            $table->unsignedBigInteger('student_id')->after('meeting_id');
             
             // Add attendance tracking columns
             $table->dateTime('join_time')->after('student_id');
